refactor(transforms): isolate mixed content parsing
